services/python/scraper/committees/requirements.txt
requests
beautifulsoup4
lxml
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
